* noted obsolete for user_options field * changed bigint to int for up_user * changed array_insert_after to wfArrayInsertAfter * doc comment for the format of XmlSelect::formatOptions * == instead of === for comparing with default in same function * moved serialize() call in saveOptions * shared logic for loadOldSearchNs
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/branches/preferences-work@49723 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ phase3/maintenance/tables.sql
@@ -86,6 +86,7 @@ CREATE TABLE /*_*/user (
   
   -- Newline-separated list of name=value defining the user
   -- preferences
+  -- Now obsolete in favour of user_properties table.
   user_options blob NOT NULL,
   
   -- This is a timestamp which is updated when a user
@@ -1269,7 +1270,7 @@ CREATE TABLE /*_*/valid_tag (
 ) /*$wgDBTableOptions*/;
 
 CREATE TABLE /*_*/user_properties(
-  up_user bigint not null,
+  up_user int not null,
   up_property varchar(255) not null,
   up_value blob
 ) /*$wgDBTableOptions*/;
